0x04. C - More functions, more nested loops

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -7,7 +7,7 @@
  * Return: square root of x
  */
 
-double _sqrt(int x)
+double _sqrt(double x)
 {
 float sqrt, t;
 sqrt = x / 2;
